Factual correction: the cities dataset only contains US cities
I don't know what the original source of the data set is. But it certainly is not a list of the worlds 1000 largest cities. As far as I can tell, the data set contains only US cities, and very few of them are among the largest 1000 cities in the world (the largest cites are mostly in Asia, where ~60% of all humans live.)